fixing useEffect dependencies on Chapter component
chapter used the setPageByEvent function from the usePageController
hook despite not including it on the dependency array. had a look at it
and noticed that since i built this hook at around 4am i didn't actually
care about following the rules of hooks and setPageByEvent depends on
the setPage function that needs to be available to be returned to
directly manipulate the page AND to be used by the setPageByEvent. i
will look into the setPage function now.